move logic to pre and post processing
