+++ src/compiler/passes/preprocess.js
@@ -1,13 +1,18 @@
-import visitor from '../visitor';
 
-let generatedWalker = visitor.build({
-  start: function(node) {
-    console.log(node);
-  }
-});
-
+let createMethodHeaders = function(name, context) {
+  name = name || context.currentIdx();
+  let f = `f${name}: function() {
+    var frag = td.${context.getTdMethodName('createDocumentFragment')}();\n`;
+  let r = `r${name}: function(c) {
+    var root = frags.frag${name} || this.f${name}();
+    root = root.cloneNode(true);\n`;
+  context.push(name, f, r);
+};
 let preprocess = function(ast, options) {
-  generatedWalker(ast);
+  let context = options.context;
+  if (context) {
+    createMethodHeaders(null, context);
+  }
 };
 
 export default preprocess;
